pyproject.toml
[tool.poetry]
name = "comiks"
version = "0.1.0"
description = ""
authors = ["@b0oml"]

[tool.poetry.dependencies]
python = "^3.7"
tabulate = "^0.8.9"
requests = "^2.27.1"
fastDamerauLevenshtein = "^1.0.7"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
comiks = 'comiks.run:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
